+++ packages/component/src/hooks/internal/useStripMarkdown.js
@@ -1,7 +1,0 @@
-import { useMemo } from 'react';
-
-import remarkStripMarkdown from '../../Utils/remarkStripMarkdown';
-
-export default function useStripMarkdown(markdown) {
-  return useMemo(() => markdown && remarkStripMarkdown(markdown).replace(/[.\s]+$/u, ''), [markdown]);
-}